refactor(src): move all files to src/
